src/main/java/br/borba/servicos/LocacaoService.java
package br.borba.servicos;

import static br.borba.utils.DataUtils.adicionarDias;

import java.util.Date;

import br.borba.entidades.Filme;
import br.borba.entidades.Locacao;
import br.borba.entidades.Usuario;
import br.borba.utils.DataUtils;

public class LocacaoService {
	
	public Locacao alugarFilme(Usuario usuario, Filme filme) {
		Locacao locacao = new Locacao();
		locacao.setFilme(filme);
		locacao.setUsuario(usuario);
		locacao.setDataLocacao(new Date());
		locacao.setValor(filme.getPrecoLocacao());

		//Entrega no dia seguinte
		Date dataEntrega = new Date();
		dataEntrega = adicionarDias(dataEntrega, 1);
		locacao.setDataRetorno(dataEntrega);
		
		//Salvando a locacao...	
		//TODO adicionar método para salvar
		
		return locacao;
	}

	public static void main(String[] args) {
		
		
		
	}
}